requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.2.1
matplotlib==3.8.3

# Optional but recommended
python-dotenv==1.0.0
plotly==5.18.0

# For future authentication
passlib==1.7.4
python-jose==3.3.0

# For potential database integration
sqlalchemy==2.0.25

authlib==1.2.1
fastapi==0.95.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
starlette==0.26.1
requests==2.28.2
python-multipart==0.0.6